Fixed layer violation
