+++ src/vs/workbench/contrib/views/browser/media/views.css
